refactored synthesizers to be more modular, passing tests
