+++ lib/src/synthesizers/utilities/utilities.dart
@@ -1,0 +1,7 @@
+// Copyright (C) 2024 Intel Corporation
+// SPDX-License-Identifier: BSD-3-Clause
+
+export 'synth_assignment.dart';
+export 'synth_logic.dart';
+export 'synth_module_definition.dart';
+export 'synth_sub_module_instantiation.dart';
